+++ stories/components/button/KButton.variant.stories.tsx
@@ -3,7 +3,7 @@ import {Container, Item} from '../common/Container';
 import {KButton} from '@/components';
 import {KButtonProps} from '@/components/button/KButton.interface';
 
-const meta: Meta<KButtonProps> = {title: 'Button', component: KButton};
+const meta: Meta<KButtonProps> = {title: 'Components/Button', component: KButton};
 
 export default meta;
